Allow abid generation in the presence of object without a suffix (#77)
When an abid model object is created without an abid archival
identifier, its suffix in the database is nil. When we attempt to create
an abid, it has to find the last identifier, and if that identifier is
nil the system cannot determine what number should come next. Therefore,
exclude nil suffix objects from consideration when determining the next
archival identifier.

Co-authored-by: Trey Pendraon <tpendragon@princeton.edu>